refactor: Merge `new_parse` module into `parse`
