+++ keyvalues-parser/src/text/mod.rs
@@ -1,3 +1,2 @@
-pub mod new_parse;
 pub mod parse;
 pub mod render;
